cambios con dani

+++ src/main/java/com/nter/final_project/presentation/dto/order/orderproduct/OrderProductInDto.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Setter
 public final class OrderProductInDto{
 
-    private int product_id;
+    private Long product_id;
     private int amount;
 }
